requirements.txt
fastapi==0.112.0
uvicorn==0.21.0
FastAPI-SQLAlchemy==0.2.1
alembic==1.10.2
pydantic==2.8.2
py-zabbix==1.1.7
Jinja2==3.1.2
python-multipart==0.0.6
aiohttp==3.8.4
asyncpg==0.27.0
aioresponses==0.7.6
